requirements-streamlit.txt
streamlit==1.41.1
google-auth-oauthlib==1.2.1
google-auth==2.37.0
pyjwt==2.10.1
extra-streamlit-components==0.1.71
python-dotenv==1.0.1
google-api-python-client==2.128.0
httpx==0.27.0
